requirements/development.txt
aniso8601==3.0.2
-e git://github.com/hersonls/djamin.git@5b5887c84341065844ab44b6cfc8ccb865ccfdb9#egg=djamin
Django==2.1.1
django-graphql-jwt==0.1.11
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
promise==2.1
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.6
